+++ src/bract/ring/inducer.clj
@@ -9,36 +9,25 @@
 
 (ns bract.ring.inducer
   (:require
-    [bract.core.config :as bc-config]
-    [bract.core.echo   :as bc-echo]
-    [bract.core.util   :as bc-util]
-    [bract.ring.config :as config]))
+    [bract.core.config :as core-config]
+    [bract.ring.config :as ring-config]))
 
 
-(defn apply-wrappers-with-context
-  "Given a context with Ring handler, look up the ring-handler wrappers `(fn [handler context]) -> handler` and apply
-  them successively to the handler, finally updating the context with the handler before returning it."
-  ([context]
-    (apply-wrappers-with-context
-      "Ring wrapper (fn [handler context])" config/cfg-wrappers-context identity context))
-  ([wrapper-type config-def f context]
-    (bc-echo/echo "Applying Ring-middleware")
-    (let [handler  (config/ctx-ring-handler context)]
-      (->> (bc-config/ctx-config context)
-        config-def
-        (bc-util/induce handler (fn [updated-handler wrapper-name]
-                                  (config/apply-wrapper-by-name
-                                    wrapper-type
-                                    (key config-def)
-                                    updated-handler
-                                    (f context)
-                                    wrapper-name)))
-        (assoc context (key config/ctx-ring-handler))))))
-
-
-(defn apply-wrappers-with-config
-  "Given a context with Ring handler, look up the ring-handler wrappers `(fn [handler config]) -> handler` and apply
-  them successively to the handler, finally updating the context with the handler before returning it."
+(defn ctx-apply-wrappers
+  "Given a context with Ring handler and Ring handler wrappers (under the context key :bract.ring/wrappers), i.e. a
+  seq of `(fn [handler context]) -> handler`, apply them in order finally updating the context with the handler."
   [context]
-  (apply-wrappers-with-context
-    "Ring wrapper (fn [handler config])" config/cfg-wrappers-config bc-config/ctx-config context))
+  (ring-config/apply-wrappers
+    context
+    (key ring-config/cfg-wrappers)
+    (ring-config/ctx-wrappers context)))
+
+
+(defn cfg-apply-wrappers
+  "Given a context with Ring handler and Ring handler wrappers (under the config key \"bract.ring.wrappers\"), i.e. a
+  seq of `(fn [handler context]) -> handler`, apply them in order finally updating the context with the handler."
+  [context]
+  (ring-config/apply-wrappers
+    context
+    (key ring-config/cfg-wrappers)
+    (-> context core-config/ctx-config ring-config/cfg-wrappers)))