+++ test/bract/ring/inducer_test.clj
@@ -10,41 +10,50 @@
 (ns bract.ring.inducer-test
   (:require
     [clojure.test :refer :all]
-    [bract.core.config :as bc-config]
+    [bract.core.config :as core-config]
+    [bract.ring.config  :as ring-config]
     [bract.ring.inducer :as inducer]))
 
 
-(defonce sample-value (volatile! 10))
+(def holder (volatile! 10))
 
 
-(defn sample-wrapper
+(defn wrapper-inc
   [handler context]
-  (vswap! sample-value (fn [^long n] (+ n 10))))
+  (vswap! holder (fn [^long n] (inc n)))
+  handler)
 
 
-(deftest test-apply-wrappers-context
+(defn wrapper-add2
+  [handler context]
+  (vswap! holder (fn [^long n] (+ n 2)))
+  handler)
+
+
+(deftest test-ctx-apply-wrappers
   (testing "happy cases"
-    (let [good-context {:bract.ring/ring-handler identity
-                        :bract.core/config {"bract.ring.wrappers.context" ['bract.ring.inducer-test/sample-wrapper]}}]
-      (vreset! sample-value 10)
-      (is (= 10 @sample-value))
-      (let [context (inducer/apply-wrappers-with-context good-context)]
-        (is (contains? context :bract.ring/ring-handler))
-        (is (contains? context :bract.core/config)))
-      (is (= 20 @sample-value))))
+    (let [context {:bract.ring/ring-handler identity
+                   :bract.ring/wrappers     [wrapper-inc wrapper-add2]}]
+      (vreset! holder 0)
+      (let [new-context (inducer/ctx-apply-wrappers context)
+            new-handler (ring-config/ctx-ring-handler new-context)]
+        (is (contains? new-context :bract.ring/wrappers))
+        (is (= :foo (new-handler :foo))))
+      (is (= 3 @holder))))
   (testing "missing/bad keys"
-    (is (thrown? IllegalArgumentException (inducer/apply-wrappers-with-context {})))))
+    (is (thrown? IllegalArgumentException (inducer/ctx-apply-wrappers {})))))
 
 
-(deftest test-apply-wrappers-config
+(deftest test-cfg-apply-wrappers
   (testing "happy cases"
-    (let [good-context {:bract.ring/ring-handler identity
-                        :bract.core/config {"bract.ring.wrappers.config" ['bract.ring.inducer-test/sample-wrapper]}}]
-      (vreset! sample-value 10)
-      (is (= 10 @sample-value))
-      (let [context (inducer/apply-wrappers-with-config good-context)]
-        (is (contains? context :bract.ring/ring-handler))
-        (is (contains? context :bract.core/config)))
-      (is (= 20 @sample-value))))
+    (let [context {:bract.ring/ring-handler identity
+                   :bract.core/config {"bract.ring.wrappers" ['bract.ring.inducer-test/wrapper-inc
+                                                              'bract.ring.inducer-test/wrapper-add2]}}]
+      (vreset! holder 0)
+      (let [new-context (inducer/cfg-apply-wrappers context)
+            new-handler (ring-config/ctx-ring-handler new-context)]
+        (is (contains? (core-config/ctx-config new-context) "bract.ring.wrappers"))
+        (is (= :foo (new-handler :foo))))
+      (is (= 3 @holder))))
   (testing "missing/bad keys"
-    (is (thrown? IllegalArgumentException (inducer/apply-wrappers-with-config {})))))
+    (is (thrown? IllegalArgumentException (inducer/cfg-apply-wrappers {})))))