+++ CHANGELOG.md
@@ -3,9 +3,25 @@
 ## TODO
 None
 
+## [WIP] 0.2.0 / 2017-June-??
+### Changed
+- Use `[bract.core "0.2.0"]`
+
+### Added
+- Config definition `bract.ring.config/ctx-wrappers`
+- Config definition `bract.ring.config/cfg-wrappers`
+- Inducer `bract.ring.inducer/ctx-apply-wrappers`
+- Inducer `bract.ring.inducer/cfg-apply-wrappers`
+
+### Removed
+- Inducer `bract.ring.inducer/apply-wrappers-with-context`
+- Inducer `bract.ring.inducer/apply-wrappers-with-config`
+- Config definition `bract.ring.config/cfg-wrappers-context`
+- Config definition `bract.ring.config/cfg-wrappers-config`
+
+
 ## 0.1.0 / 2017-April-25
 ### Added
 - Development/test support for Ring handlers
 - Bract inducer fn to apply configured wrappers (context)
 - Bract inducer fn to apply configured wrappers (config)
-